fix: refresh load
